resource_servers/cc_middleware/requirements.txt
Flask==3.0.2
PyKCS11
flask-cors
cryptography
pyjwt